requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

alembic==1.14.1
requests==2.32.3
SQLAlchemy==2.0.37
SQLAlchemy-Utils==0.41.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9